Corrigi bugs da aula de ncurses 1

+++ examples/ncurses-1/04-oo_model/oo_model.hpp
@@ -34,7 +34,7 @@ class Fisica {
   public:
     Fisica(ListaDeCorpos *ldc);
     void add_corpo(Corpo *c);
-    void update(uint64_t deltaT);
+    void update(float deltaT);
 };
 
 class Tela {
